+++ spec/lib/ruby_lokalise_api/oauth2/auth_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe RubyLokaliseApi::OAuth2::Auth do
   describe '#auth' do
     it 'returns auth code' do
       uri = auth_client.auth scope: 'read_projects'
-      expect(uri).to include(described_class::BASE_URL.to_s)
+      expect(uri).to include(auth_client.base_url.to_s)
       expect(uri).to include('auth?client_id=')
       expect(uri).to include('&scope=read_projects')
     end
@@ -12,21 +12,21 @@ RSpec.describe RubyLokaliseApi::OAuth2::Auth do
     it 'allows to pass an array of scopes' do
       scopes = %w[write_projects write_team_groups write_tasks]
       uri = auth_client.auth scope: scopes
-      expect(uri).to include(described_class::BASE_URL.to_s)
+      expect(uri).to include(auth_client.base_url.to_s)
       expect(uri).to include('auth?client_id=')
       expect(uri).to include("&scope=#{scopes.join('+')}")
     end
 
     it 'allows to set redirect_uri' do
       uri = auth_client.auth scope: 'read_projects', redirect_uri: 'http://example.com/callback'
-      expect(uri).to include(described_class::BASE_URL.to_s)
+      expect(uri).to include(auth_client.base_url.to_s)
       expect(uri).to include('example.com%2Fcallback')
     end
 
     it 'allows to set state' do
       state = '123abc'
       uri = auth_client.auth scope: 'read_projects', state: state
-      expect(uri).to include(described_class::BASE_URL.to_s)
+      expect(uri).to include(auth_client.base_url.to_s)
       expect(uri).to include(state)
     end
   end
@@ -36,10 +36,12 @@ RSpec.describe RubyLokaliseApi::OAuth2::Auth do
       resp = VCR.use_cassette('oauth2/token') do
         auth_client.token ENV.fetch('OAUTH2_CODE', nil)
       end
-      expect(resp['access_token']).not_to be_nil
-      expect(resp['refresh_token']).not_to be_nil
-      expect(resp['expires_in']).to eq(3600)
-      expect(resp['token_type']).to eq('Bearer')
+
+      expect(resp).to be_an_instance_of(RubyLokaliseApi::OAuth2::Token)
+      expect(resp.access_token).not_to be_nil
+      expect(resp.refresh_token).not_to be_nil
+      expect(resp.expires_in).to eq(3600)
+      expect(resp.token_type).to eq('Bearer')
     end
 
     it 'raises an error when the code is invalid' do
@@ -56,10 +58,12 @@ RSpec.describe RubyLokaliseApi::OAuth2::Auth do
       resp = VCR.use_cassette('oauth2/refresh') do
         auth_client.refresh ENV.fetch('OAUTH2_REFRESH_TOKEN', nil)
       end
-      expect(resp['access_token']).not_to be_nil
-      expect(resp['refresh_token']).to be_nil
-      expect(resp['expires_in']).to eq(3600)
-      expect(resp['token_type']).to eq('Bearer')
+
+      expect(resp).to be_an_instance_of(RubyLokaliseApi::OAuth2::Refresh)
+      expect(resp.access_token).not_to be_nil
+      expect(resp.scope).to eq('write_projects write_team_groups write_tasks')
+      expect(resp.expires_in).to eq(3600)
+      expect(resp.token_type).to eq('Bearer')
     end
 
     it 'raises an error when the token is invalid' do